refactor: Clearly name all decorators
This is done because when exploring the code, it's not always clear
what the purpose of the class is unless it's clearly named
